move validations to the link helper
